+++ src/views/form/designer/RightPanel/index.vue
@@ -177,6 +177,11 @@
       </div>
       <div v-if="hasOwnProperty(element, 'code')">
         <el-divider>代码</el-divider>
+        <FullScreen>
+          <div style="height: 85vh; border: 1px solid #ebeef5;">
+            <MonacoEditor v-model="element.code" language="html" />
+          </div>
+        </FullScreen>
         <el-input
           type="textarea"
           v-model="element.code"